+++ R/drive_mv.R
@@ -22,7 +22,8 @@
 #' @export
 #' @examplesIf drive_has_token()
 #' # create a file to move
-#' file <- drive_upload(drive_example("chicken.txt"), "chicken-mv.txt")
+#' file <- drive_example_remote("chicken.txt") %>%
+#'   drive_cp("chicken-mv.txt")
 #'
 #' # rename it, but leave in current folder (root folder, in this case)
 #' file <- drive_mv(file, "chicken-mv-renamed.txt")