+++ R/drive_upload.R
@@ -34,17 +34,15 @@
 #' @export
 #' @examplesIf drive_has_token()
 #' # upload a csv file
-#' chicken_csv <- drive_upload(
-#'   drive_example("chicken.csv"),
-#'   "chicken-upload.csv"
-#' )
+#' chicken_csv <- drive_example_local("chicken.csv") %>%
+#'   drive_upload("chicken-upload.csv")
 #'
 #' # or convert it to a Google Sheet
-#' chicken_sheet <- drive_upload(
-#'   drive_example("chicken.csv"),
-#'   name = "chicken-sheet-upload.csv",
-#'   type = "spreadsheet"
-#' )
+#' chicken_sheet <- drive_example_local("chicken.csv") %>%
+#'   drive_upload(
+#'     name = "chicken-sheet-upload.csv",
+#'     type = "spreadsheet"
+#'   )
 #'
 #' # check out the new Sheet!
 #' drive_browse(chicken_sheet)
@@ -53,10 +51,8 @@
 #' drive_find("chicken.*upload") %>% drive_rm()
 #'
 #' # Upload a file and, at the same time, star it
-#' chicken <- drive_upload(
-#'   drive_example("chicken.jpg"),
-#'   starred = "true"
-#' )
+#' chicken <- drive_example_local("chicken.jpg") %>%
+#'   drive_upload(starred = "true")
 #'
 #' # Is is really starred? YES
 #' purrr::pluck(chicken, "drive_resource", 1, "starred")
@@ -67,18 +63,18 @@
 #' # `overwrite = FALSE` errors if something already exists at target filepath
 #' # THIS WILL ERROR!
 #' drive_create("name-squatter")
-#' drive_upload(
-#'   drive_example("chicken.jpg"),
-#'   name = "name-squatter",
-#'   overwrite = FALSE
-#' )
+#' drive_example_local("chicken.jpg") %>%
+#'   drive_upload(
+#'     name = "name-squatter",
+#'     overwrite = FALSE
+#'   )
 #'
 #' # `overwrite = TRUE` moves the existing item to trash, then proceeds
-#' chicken <- drive_upload(
-#'   drive_example("chicken.jpg"),
-#'   name = "name-squatter",
-#'   overwrite = TRUE
-#' )
+#' chicken <- drive_example_local("chicken.jpg") %>%
+#'   drive_upload(
+#'     name = "name-squatter",
+#'     overwrite = TRUE
+#'   )
 #'
 #' # Clean up
 #' drive_rm(chicken)